# docs/uapoly-openapi.yaml
openapi: 3.1.0
...
info:
  title: UApoly
  version: '1.0'
  description: The API specification for interacting with the UApoly backend.
  x-logo:
    url: /docs/Logo.webp
servers:
  - url: 'http://localhost:{port}'
    description: Local instance
    variables:
      port:
        default: '3000'
        description: The port where the local instance of the UApoly backend is running.
    x-last-modified: 1680690602604
paths:
  /user/login:
    summary: Log-in an user.
    post:
      requestBody:
        content:
          application/json:
            schema:
              description: 'A JSON object with the following attributes :'
              required:
                - login
                - password
              type: object
              properties:
                login:
                  description: The name of the account
                  type: string
                  pattern: '^([a-zA-Z]|[0-9]|-|_){4,15}$'
                password:
                  description: The password of the account
                  type: string
                  minLength: 6
              externalDocs:
                url: ''
                description: ''
            example:
              login: tchm
              password: password
        required: true
        description: ''
      tags: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
              example:
                token: USER_TOKEN
          description: Successful authentification.
          x-last-modified: 1680685519518
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                message: Missing login or password
          description: Missing login or password
          x-last-modified: 1680685718584
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                message: Invalid password
          description: Authentification failure.
          x-last-modified: 1680685787933
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                message: User not found
          description: Invalid login.
          x-last-modified: 1680685662119
      operationId: /user/login
      summary: Logs in an user
      description: 'Get a security token for a user, effectively logging them in.'
      security: []
    x-last-modified: 1680684450769
  /user/register:
    summary: Create a new account
    post:
      requestBody:
        content:
          application/json:
            schema:
              description: 'A JSON object with the following attributes :'
              required:
                - login
                - password
                - email
              type: object
              properties:
                login:
                  description: The name of the account
                  type: string
                  pattern: '^([a-zA-Z]|[0-9]|-|_){4,15}$'
                password:
                  description: The password of the account
                  type: string
                  minLength: 6
                email:
                  description: The email of the new account
                  type: string
              externalDocs:
                url: ''
                description: ''
            example:
              login: tchm
              password: password
              email: tchm@example.org
        required: true
        description: ''
      tags: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
              example:
                token: USER_TOKEN
          description: Account created
          x-last-modified: 1680686689900
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                message: 'Missing login, password or email'
          description: 'Missing or invalid login, password or email'
          x-last-modified: 1680686490021
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                message: Login already taken
          description: Login already taken
          x-last-modified: 1680686627305
      operationId: /user/register
      summary: Create a new account
      description: 'Create a new account. Also returns a security token for the newly created account, logging them in.'
      security: []
    x-last-modified: 1680686045560
  /user/me:
    summary: Get details about the current account
    get:
      tags: []
      responses:
        '200':
          content:
            application/json:
              schema:
                description: 'A JSON object with the following attributes :'
                required:
                  - login
                  - email
                type: object
                properties:
                  login:
                    description: The name of the current account
                    type: string
                  email:
                    description: The email associated with the current account
                    type: string
                externalDocs:
                  description: ''
                  url: ''
              example:
                login: tchm
                email: tchm@example.org
          description: ''
          x-last-modified: 1680687881073
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                message: User not found
          description: User not found
          x-last-modified: 1680687810041
      operationId: /user/me
      summary: Get details about the current account.
      description: Returns account details for the currently logged in user.
    x-last-modified: 1680687612749
...
  /user/search:
    summary: Search for usernames.
    description: Search for usernames.
    post:
      requestBody:
        content:
          application/json:
            schema:
              type: object
              externalDocs:
                description: ''
                url: ''
              properties:
                login:
                  description: The search query
                  type: string
                  minLength: 1
                page:
                  description: 'If there are more than 10 results, select the page to get. Defaults to 1 if unspecified'
                  minimum: 1
                  type: integer
              required:
                - login
            example:
              login: h
              page: 1
            examples:
              Example:
                value:
                  login: tch
                  page: 1
        required: true
        description: 'A JSON object with :'
      tags: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountList'
              example:
                - login: tchm
                - login: Gyrehio
              examples:
                Example:
                  value:
                    - login: tchm
          description: A JSON array containing the results.
          x-last-modified: 1680771346183
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                message: Invalid page
          description: Missing login or invalid page.
          x-last-modified: 1680771629794
      operationId: /user/search
      summary: Search for usernames
      description: Search for usernames
    x-last-modified: 1680771177197
...
components:
  schemas:
...
    Token:
      description: 'An object containing a security token, allowing to do actions as the user associated with it.'
      required:
        - token
      type: object
      properties:
        token:
          description: A security token
          type: string
      x-last-modified: 1680690400975
    AccountName:
      description: An object representing an account name
      required:
        - login
      type: object
      properties:
        login:
          description: The name of an account
          type: string
      x-last-modified: 1680691637662
    Message:
      description: 'Represents an message, either a success or an error'
      required:
        - message
      type: object
      properties:
        message:
          description: The message
          type: string
      x-last-modified: 1680692391186
    AccountList:
      description: A list of accounts with their logins.
      type: array
      items:
        $ref: '#/components/schemas/AccountName'
      x-last-modified: 1680771126932
...
  securitySchemes:
    JWTAuth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: |-
        Example: 
        > Authorization: Bearer <token>
      x-last-modified: 1680684271664
...
security:
  - JWTAuth: []
...